added modal to delete location

+++ src/components/forms/NewLocation.jsx
@@ -7,6 +7,7 @@ import {
     updateLocationByLocationId,
 } from "../../services/locationService";
 import {
+    AlertDialog,
     Button,
     Card,
     Container,
@@ -124,15 +125,41 @@ export const NewLocation = () => {
                         >
                             Save Location
                         </Button>
-                        <Button
-                            color="red"
-                            m="2"
-                            onClick={() => {
-                                handleDeleteLocation();
-                            }}
-                        >
-                            Delete Location
-                        </Button>
+                        <AlertDialog.Root>
+                            <AlertDialog.Trigger>
+                                <Button m="2" color="red">
+                                    Delete Location
+                                </Button>
+                            </AlertDialog.Trigger>
+                            <AlertDialog.Content size="1" maxWidth="300px">
+                                <AlertDialog.Title>
+                                    Delete Location
+                                </AlertDialog.Title>
+                                <AlertDialog.Description size="2">
+                                    Are you sure? This Location, its projects,
+                                    tasks and items will be deleted.
+                                </AlertDialog.Description>
+
+                                <Flex gap="3" mt="4" justify="end">
+                                    <AlertDialog.Cancel>
+                                        <Button variant="soft" color="gray">
+                                            Cancel
+                                        </Button>
+                                    </AlertDialog.Cancel>
+                                    <AlertDialog.Action>
+                                        <Button
+                                            variant="solid"
+                                            color="red"
+                                            onClick={() => {
+                                                handleDeleteLocation();
+                                            }}
+                                        >
+                                            Delete Location
+                                        </Button>
+                                    </AlertDialog.Action>
+                                </Flex>
+                            </AlertDialog.Content>
+                        </AlertDialog.Root>
                     </Flex>
                 </Section>
             </Card>
